transaction checking if data has meaning and if date is already taken

+++ festou/Festou/api/manager.py
@@ -4,6 +4,7 @@ from rest_framework.response import Response
 from rest_framework import status
 from django.http import JsonResponse
 from datetime import datetime, timedelta
+from django.utils import timezone
 import hashlib
 
 def create_user(self, request): 
@@ -51,9 +52,12 @@ def create_user(self, request):
 def edit_user(self, request, user_id):
     try:
         user = User.objects.get(pk=user_id)
+        # Verificar se é o usuário 
+        if user_id != request.user.id:
+            return Response({'message': 'You are not the owner of this account.'}, status=status.HTTP_403_FORBIDDEN)   
         serializer = self.serializer_class(user, data=request.data)
         if serializer.is_valid():
-            queryset = User.objects.filter(email = user.email)
+            queryset = User.objects.filter(email = email)
             if queryset.exists():
                 return Response({'description': 'CPF or Email already linked to an existing account. Please try again.'}, status=status.HTTP_400_BAD_REQUEST)
             serializer.save()
@@ -103,6 +107,9 @@ def create_place(self, request):
 def edit_place(self, request, place_id):
     try:
         place = Place.objects.get(pk=place_id)
+        # Verificar se o usuário é o proprietário do lugar
+        if place.id_owner != request.user.id:
+            return Response({'message': 'You are not the owner of this place.'}, status=status.HTTP_403_FORBIDDEN)   
         serializer = self.serializer_class(place, data=request.data)
         if serializer.is_valid():
             queryset = Place.objects.filter(location = place.location)
@@ -118,6 +125,9 @@ def edit_place(self, request, place_id):
 def delete_place(self, request, place_id):
     try:
         place = Place.objects.get(pk=place_id)
+        # Verificar se o usuário é o proprietário do lugar
+        if place.id_owner != request.user.id:
+            return Response({'message': 'You are not the owner of this place.'}, status=status.HTTP_403_FORBIDDEN)
         place.delete()
         return Response({'message': 'Place deleted successfully.'}, status=status.HTTP_204_NO_CONTENT)
     except Place.DoesNotExist:
@@ -132,6 +142,18 @@ def create_transaction(self, request):
         id_place = serializer.validated_data.get("id_place")
         initial_date = serializer.validated_data.get("initial_date")
         final_date = serializer.validated_data.get("final_date")
+
+        #vendo se faz sentido
+        if initial_date < timezone.now().date() or final_date < initial_date:
+            return Response({'description': 'Invalid date range...'}, status=status.HTTP_400_BAD_REQUEST)
+        overlapping_transactions = Transaction.objects.filter(
+            id_place=id_place,
+            initial_date__lte=final_date,
+            final_date__gte=initial_date,
+        )
+        if overlapping_transactions.exists():
+            return Response({'description': 'Date conflict with existing transactions...'}, status=status.HTTP_400_BAD_REQUEST)
+
         days_to_subtract = 7 #Quantidade de dias antes do evento que será liberado o dinheiro 
         payday = initial_date - timedelta(days=days_to_subtract)
         atual_place = Place.objects.get(pk = id_place)
